apps: separate volume allocation methods from standalone funcs
Copy functions out of  volume_entry_allocate.go into a new file
brick_allocate.go, so that volume_entry_allocate.go only contains
methods on the VolumeEntry struct and brick_allocate.go only
contains the standalone utility methods.

The expectation is that more code will be made into common
utility functions to support future allocator enhancements
and it needs a better home than sleeping on VolumeEntry's couch.
